feat: JB_27-APR-24 Improvements to user service - added open api v3 doc in resources - added health endpoint - corrections to package name and versions
