+++ pom.xml
@@ -8,10 +8,10 @@
         <version>3.2.2</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
-    <groupId>com.jesslabs</groupId>
-    <artifactId>usermanager</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <name>usermanager</name>
+    <groupId>com.jesslabs.user_management</groupId>
+    <artifactId>user-management</artifactId>
+    <version>1.0.0</version>
+    <name>user_management</name>
     <description>User Manager App</description>
     <properties>
         <java.version>21</java.version>
@@ -34,6 +34,10 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-actuator</artifactId>
         </dependency>
         <dependency>
             <groupId>org.postgresql</groupId>
