+++ src/main/java/com/jesslabs/user_management/contoller/HealthController.java
@@ -1,0 +1,18 @@
+package com.jesslabs.user_management.contoller;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@OpenAPIDefinition(info = @Info(title = "Health controller"), tags = @Tag(name = "Health Controller"))
+public class HealthController {
+    @GetMapping("/health")
+    public ResponseEntity<String> healthCheck() {
+        return ResponseEntity.status(HttpStatus.OK).body("Success");
+    }
+}
