+++ src/main/java/com/jesslabs/user_management/dto/GetUsersPagedDTO.java
@@ -1,0 +1,15 @@
+package com.jesslabs.user_management.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class GetUsersPagedDTO {
+    private List<GetUsersDTO> users;
+    private PageDTO page;
+}
