+++ src/main/java/com/jesslabs/user_management/dto/PageDTO.java
@@ -1,0 +1,15 @@
+package com.jesslabs.user_management.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class PageDTO {
+    private int totalPages;
+    private int currentPageNo;
+    private int currentPageSize;
+    private Long total;
+}
