+++ src/main/java/com/jesslabs/user_management/dto/UpdateUserRequestDTO.java
@@ -1,0 +1,16 @@
+package com.jesslabs.user_management.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class UpdateUserRequestDTO {
+    private String name;
+    private String password;
+    private String email;
+    private String role;
+    private String phone;
+}
