+++ src/main/java/com/jesslabs/user_management/exception/BadRequestException.java
@@ -1,0 +1,7 @@
+package com.jesslabs.user_management.exception;
+
+public class BadRequestException extends RuntimeException {
+    public BadRequestException(String message) {
+        super(message);
+    }
+}
