+++ src/main/java/com/jesslabs/user_management/exception/ConflictException.java
@@ -1,0 +1,7 @@
+package com.jesslabs.user_management.exception;
+
+public class ConflictException extends RuntimeException {
+    public ConflictException(String message) {
+        super(message);
+    }
+}
