+++ src/main/java/com/jesslabs/user_management/exception/HttpException.java
@@ -1,0 +1,15 @@
+package com.jesslabs.user_management.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class HttpException extends RuntimeException {
+    private final HttpStatus status;
+
+    public HttpException(String message, HttpStatus status) {
+        super(message);
+        this.status = status;
+    }
+
+}
