+++ src/main/java/com/jesslabs/user_management/exception/InternalServerException.java
@@ -1,0 +1,7 @@
+package com.jesslabs.user_management.exception;
+
+public class InternalServerException extends RuntimeException {
+    public InternalServerException(String message) {
+        super(message);
+    }
+}
