+++ src/main/java/com/jesslabs/user_management/exception/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package com.jesslabs.user_management.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
