--- src/main/java/com/jesslabs/usermanager/dto/GetUsersPagedDTO.java
@@ -1,15 +1,0 @@
-package com.jesslabs.usermanager.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class GetUsersPagedDTO {
-    private List<GetUsersDTO> users;
-    private PageDTO page;
-}
