--- src/main/java/com/jesslabs/usermanager/dto/PageDTO.java
@@ -1,15 +1,0 @@
-package com.jesslabs.usermanager.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class PageDTO {
-    private int totalPages;
-    private int currentPageNo;
-    private int currentPageSize;
-    private Long total;
-}
