--- src/main/java/com/jesslabs/usermanager/dto/UpdateUserRequestDTO.java
@@ -1,16 +1,0 @@
-package com.jesslabs.usermanager.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UpdateUserRequestDTO {
-    private String name;
-    private String password;
-    private String email;
-    private String role;
-    private String phone;
-}
