--- src/main/java/com/jesslabs/usermanager/dto/UpdateUserResponseDTO.java
@@ -1,17 +1,0 @@
-package com.jesslabs.usermanager.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UpdateUserResponseDTO {
-    private Long id;
-    private String name;
-    private String username;
-    private String email;
-    private String role;
-    private String phone;
-}
