--- src/main/java/com/jesslabs/usermanager/exception/BadRequestException.java
@@ -1,7 +1,0 @@
-package com.jesslabs.usermanager.exception;
-
-public class BadRequestException extends RuntimeException {
-    public BadRequestException(String message) {
-        super(message);
-    }
-}
