--- src/main/java/com/jesslabs/usermanager/exception/ConflictException.java
@@ -1,7 +1,0 @@
-package com.jesslabs.usermanager.exception;
-
-public class ConflictException extends RuntimeException {
-    public ConflictException(String message) {
-        super(message);
-    }
-}
