--- src/main/java/com/jesslabs/usermanager/exception/HttpException.java
@@ -1,15 +1,0 @@
-package com.jesslabs.usermanager.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class HttpException extends RuntimeException {
-    private final HttpStatus status;
-
-    public HttpException(String message, HttpStatus status) {
-        super(message);
-        this.status = status;
-    }
-
-}
