--- src/main/java/com/jesslabs/usermanager/exception/InternalServerException.java
@@ -1,7 +1,0 @@
-package com.jesslabs.usermanager.exception;
-
-public class InternalServerException extends RuntimeException {
-    public InternalServerException(String message) {
-        super(message);
-    }
-}
