--- src/main/java/com/jesslabs/usermanager/exception/ResourceNotFoundException.java
@@ -1,7 +1,0 @@
-package com.jesslabs.usermanager.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
