--- src/main/java/com/jesslabs/usermanager/model/User.java
@@ -1,39 +1,0 @@
-package com.jesslabs.usermanager.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity(name = "app_user")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false, unique = true)
-    private String username;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String role;
-
-    @Column(nullable = false)
-    private String phone;
-
-    @Column(nullable = false)
-    private boolean isDiscarded;
-
-}
