--- src/main/java/com/jesslabs/usermanager/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.jesslabs.usermanager.repository;
-
-import com.jesslabs.usermanager.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    boolean existsUserByUsername(String username);
-
-    Optional<User> findUserByUsername(String username);
-
-
-}
