+++ src/main/resources/application.properties
@@ -5,3 +5,6 @@ spring.datasource.password=12345
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+
+endpoints.info.enabled=true
+info.app.version=@project.version@
